refactor(api-core): rename request.py to requests.py for consistency
