+++ packages/google-api-core/google/api_core/gapic_v1/__init__.py
@@ -25,10 +25,10 @@ _has_grpc = importlib.util.find_spec("grpc") is not None
 # Older Python versions safely ignore this variable.
 __lazy_modules__: Set[str] = {
     "google.api_core.gapic_v1.client_info",
-    "google.api_core.gapic_v1.request",
+    "google.api_core.gapic_v1.requests",
     "google.api_core.gapic_v1.routing_header",
 }
-__all__ = ["client_info", "request", "routing_header"]
+__all__ = ["client_info", "requests", "routing_header"]
 
 if _has_grpc:
     __lazy_modules__.update(
@@ -42,7 +42,7 @@ if _has_grpc:
 
 from google.api_core.gapic_v1 import (  # noqa: E402
     client_info,
-    request,
+    requests,
     routing_header,
 )
