+++ packages/google-api-core/tests/unit/gapic/test_requests.py
@@ -17,7 +17,7 @@ import re
 
 import pytest
 
-from google.api_core.gapic_v1.request import setup_request_id
+from google.api_core.gapic_v1.requests import setup_request_id
 
 
 # --- Mock Request Helper Classes ---
